fix: Disable Red Hat YAML telemetry prompts in Codespaces
When opening the project in Codespaces, users were prompted to allow
telemetry data to be sent to Red Hat from the YAML extension. This
creates an unnecessary interruption to the development workflow.

This commit adds 'redhat.telemetry.enabled: false' to the VS Code
workspace settings to disable these prompts by default.

Fixes #401